Repair clang compilation

+++ src/interpreter/cpu.cpp
@@ -1362,9 +1362,9 @@ namespace x64 {
     void Cpu::exec(const Fstp<ST>& ins) { x87fpu_.set(ins.dst, x87fpu_.st(ST::ST0)); x87fpu_.pop(); }
     void Cpu::exec(const Fstp<M80>& ins) { set(resolve(ins.dst), x87fpu_.st(ST::ST0)); x87fpu_.pop(); }
 
-    void Cpu::exec(const Fistp<M16>& ins) { set(resolve(ins.dst), f80::castToI16(x87fpu_.st(ST::ST0))); x87fpu_.pop(); }
-    void Cpu::exec(const Fistp<M32>& ins) { set(resolve(ins.dst), f80::castToI32(x87fpu_.st(ST::ST0))); x87fpu_.pop(); }
-    void Cpu::exec(const Fistp<M64>& ins) { set(resolve(ins.dst), f80::castToI64(x87fpu_.st(ST::ST0))); x87fpu_.pop(); }
+    void Cpu::exec(const Fistp<M16>& ins) { set(resolve(ins.dst), (u16)f80::castToI16(x87fpu_.st(ST::ST0))); x87fpu_.pop(); }
+    void Cpu::exec(const Fistp<M32>& ins) { set(resolve(ins.dst), (u32)f80::castToI32(x87fpu_.st(ST::ST0))); x87fpu_.pop(); }
+    void Cpu::exec(const Fistp<M64>& ins) { set(resolve(ins.dst), (u64)f80::castToI64(x87fpu_.st(ST::ST0))); x87fpu_.pop(); }
 
     void Cpu::exec(const Fxch<ST>& ins) {
         f80 src = x87fpu_.st(ins.src);
